refactor: split admin handler
